Rename LinkdingBookmarkTabSettingsView to LinkdingBookmarkTabFilterView Simplify settings view and apply changes immediately
